requirements.txt
Django>=4.0,<4.1
django-crispy-forms
django-weasyprint
django-settings-export
django-extra-views
django-multiselectfield
python-magic
